refactor(inference): extract pure-ops library (PR 1 of #508)
Move the stateless inference primitives into a single namespace
``sleap_nn.inference.ops`` so subsequent layer/backend PRs (PR 4-7) can
import without pulling in Lightning, sleap-io, or the predictor stack.

Layout:

  sleap_nn/inference/ops/
    peaks.py     find_*_peaks{,_rough}, integral_regression,
                 morphological_dilation
    paf.py       PAFScorer + helpers (whole-file move)
    coord.py     NEW — undo_stride / undo_input_scale / undo_eff_scale /
                 add_crop_offset / apply_input_scale, the single source of
                 truth for the coordinate-reverse ladder. Identity ops
                 short-circuit so the no-op cases pay zero ops.
    filters.py   filter_by_node_count / _confidence /
                 filter_overlapping_instances + private NMS helpers
                 (whole-file move from postprocessing.py)
    identity.py  classify_peaks_from_maps + group_class_peaks
                 (whole-file move)
    crops.py     crop_bboxes (extracted from peak_finding.py) and a
                 re-export of make_centered_bboxes from the data layer

The original modules (peak_finding.py, paf_grouping.py, postprocessing.py,
identity.py) become thin re-export shims so every existing caller keeps
working unchanged. They are scheduled for deletion in #519.

Tests:

  tests/inference/ops/test_coord.py        17 unit tests for the new ops,
                                           including a full forward/back
                                           ladder round-trip and an MPS
                                           parametrization
  tests/inference/ops/test_onnx_export.py  Smoke tests that every op in
                                           peaks/coord/crops still exports
                                           cleanly to ONNX opset 16. Two
                                           pre-existing exporter limitations
                                           (Tensor.unfold,
                                           boolean-mask index_put_) are
                                           pinned as strict xfails so PR 5
                                           (#513) is reminded to fix them.

Parity: ``RUN_GOLDEN_REGEN_CHECK=1 pytest tests/inference/test_parity_goldens.py``
passes on all six specs — byte-for-byte against the PR 0 goldens. Existing
inference suite (149 tests) passes through the shims unchanged.
