+++ tests/inference/ops/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for ``sleap_nn.inference.ops``."""
